[vulkan] Fix warnings: static_cast, remove unused
ghstack-source-id: 234686d1c3783092c7a9cd4040a0f04b5f46b024
Pull Request resolved: https://github.com/pytorch/pytorch/pull/42195

+++ aten/src/ATen/native/vulkan/Vulkan.cpp
@@ -992,7 +992,8 @@ void copy_buffer_to_image(const VBuffer& buffer, VImage& image) {
     int32_t w;
     int32_t h;
   };
-  const ConstBlock constBlock{image.w(), image.h()};
+  const ConstBlock constBlock{static_cast<int32_t>(image.w()),
+                              static_cast<int32_t>(image.h())};
   VBuffer constBuffer =
       makeUniformConstBuffer(&constBlock, sizeof(constBlock));
 
@@ -1053,7 +1054,8 @@ void copy_image_to_buffer(
     int32_t w;
     int32_t h;
   };
-  const ConstBlock constBlock{image.w(), image.h()};
+  const ConstBlock constBlock{static_cast<int32_t>(image.w()),
+                              static_cast<int32_t>(image.h())};
   VBuffer constBuffer =
       makeUniformConstBuffer(&constBlock, sizeof(constBlock));
